refactor and added tests
